[exercise] исправлена ошибка, с тем, что пользователь может удалить упражнение, которое он не создавал

+++ sport-app-backend/src/exercises/exercises.controller.ts
@@ -84,7 +84,7 @@ export class ExercisesController {
   }
 
   @Delete(':id')
-  async del(@Param('id', ParseIntPipe) id) {
-    return await this.exercisesService.deleteExercise({ id: id });
+  async del(@Param('id', ParseIntPipe) id, @User() user) {
+    return await this.exercisesService.deleteExercise({ id: id }, user.userId);
   }
 }
